The bios transition freq is in 10s kHz - so x10

+++ src/nv_type.h
@@ -316,7 +316,7 @@ typedef struct {
 		Bool dual_link;
 		Bool if_is_18bit;
 		Bool BITbit1;
-		uint16_t duallink_transition_clk;
+		int duallink_transition_clk;
 	} fp;
 
 	struct {
